Better handling of insertion indexes
This ensures we do the best possible job to reinsert the new instance at the appropriate index inside its parent. There is some error handling with the indexes to make sure the index isn't greater than number of children (which can happen if you are consolidating multiple elements down to a single component). If there is no children, we insert it at 0. If there are but less than index, we insert it at one less  than the length of the children array.